+++ crates/barracuda/src/shaders/misc/argsort_f64.wgsl
@@ -1,4 +1,4 @@
-// ArgSort - Return indices that sort the tensor (GPU parallel)
+// ArgSort_f64.wgsl — Return indices that sort the tensor (f64 canonical)
 // Returns indices that would sort the input tensor
 //
 // Algorithm: Parallel bitonic sort with index tracking
@@ -12,7 +12,7 @@ struct Params {
 }
 
 @group(0) @binding(0) var<uniform> params: Params;
-@group(0) @binding(1) var<storage, read_write> values: array<f32>;   // Values being sorted
+@group(0) @binding(1) var<storage, read_write> values: array<f64>;   // Values being sorted
 @group(0) @binding(2) var<storage, read_write> indices: array<u32>;  // Indices being sorted
 
 @compute @workgroup_size(256)
@@ -25,31 +25,31 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
     // Bitonic sort: compute partner index
     let step_size = 1u << params.step;
     let stage_size = 1u << params.stage;
-    
+
     let block_idx = idx / step_size;
     let within_block = idx % step_size;
     let partner_idx = (block_idx * 2u * step_size) + (2u * step_size - 1u - within_block);
     let my_idx = (block_idx * 2u * step_size) + within_block;
-    
+
     if (partner_idx < params.size && my_idx < params.size) {
         let val1 = values[my_idx];
         let val2 = values[partner_idx];
-        
+
         // Determine sort direction
         let ascending = ((my_idx / stage_size) % 2u) == 0u;
         var should_swap = false;
-        
+
         if (params.descending == 0u) {
             should_swap = (ascending && val1 > val2) || (!ascending && val1 < val2);
         } else {
             should_swap = (ascending && val1 < val2) || (!ascending && val1 > val2);
         }
-        
+
         if (should_swap) {
             // Swap values
             values[my_idx] = val2;
             values[partner_idx] = val1;
-            
+
             // Swap indices
             let temp_idx = indices[my_idx];
             indices[my_idx] = indices[partner_idx];
